requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
pydantic==2.1.1
python-dotenv==1.0.0
requests==2.31.0
crewai==0.28.7
python-multipart==0.0.6